Move bukkit-legacy into bukkit
I don't see any reason for keeping bukkit-legacy as the only relevant change is getTypeId() which can be accessed through getType().getId(). also renames some remaining packages to the new protocol name format.
