--- bukkit-legacy/build.gradle.kts
@@ -1,8 +1,0 @@
-dependencies {
-    compileOnlyApi(projects.viaversionCommon)
-    compileOnly(libs.legacyBukkit) {
-        exclude("junit", "junit")
-        exclude("com.google.code.gson", "gson")
-        exclude("javax.persistence", "persistence-api")
-    }
-}
