+++ bukkit/src/main/java/com/viaversion/viaversion/ViaVersionPlugin.java
@@ -26,7 +26,7 @@ import com.viaversion.viaversion.api.platform.UnsupportedSoftware;
 import com.viaversion.viaversion.api.platform.ViaPlatform;
 import com.viaversion.viaversion.bukkit.commands.BukkitCommandHandler;
 import com.viaversion.viaversion.bukkit.commands.BukkitCommandSender;
-import com.viaversion.viaversion.bukkit.listeners.JoinListener;
+import com.viaversion.viaversion.bukkit.listeners.base.JoinListener;
 import com.viaversion.viaversion.bukkit.platform.BukkitViaAPI;
 import com.viaversion.viaversion.bukkit.platform.BukkitViaConfig;
 import com.viaversion.viaversion.bukkit.platform.BukkitViaInjector;
