+++ bukkit/src/main/java/com/viaversion/viaversion/bukkit/listeners/base/JoinListener.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.viaversion.viaversion.bukkit.listeners;
+package com.viaversion.viaversion.bukkit.listeners.base;
 
 import com.viaversion.viaversion.api.Via;
 import com.viaversion.viaversion.api.connection.ProtocolInfo;
