+++ bukkit/src/main/java/com/viaversion/viaversion/bukkit/listeners/v1_8to1_9/BlockListener.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.viaversion.viaversion.bukkit.listeners.protocol1_9to1_8;
+package com.viaversion.viaversion.bukkit.listeners.v1_8to1_9;
 
 import com.viaversion.viaversion.api.minecraft.BlockPosition;
 import com.viaversion.viaversion.bukkit.listeners.ViaBukkitListener;
