+++ bukkit/src/main/java/com/viaversion/viaversion/bukkit/listeners/v1_8to1_9/HandItemCache.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.viaversion.viaversion.bukkit.listeners.protocol1_9to1_8;
+package com.viaversion.viaversion.bukkit.listeners.v1_8to1_9;
 
 import com.viaversion.viaversion.api.minecraft.item.DataItem;
 import com.viaversion.viaversion.api.minecraft.item.Item;
@@ -53,6 +53,6 @@ public class HandItemCache extends BukkitRunnable {
 
     public static Item convert(ItemStack itemInHand) {
         if (itemInHand == null) return new DataItem(0, (byte) 0, (short) 0, null);
-        return new DataItem(itemInHand.getTypeId(), (byte) itemInHand.getAmount(), itemInHand.getDurability(), null);
+        return new DataItem(itemInHand.getType().getId(), (byte) itemInHand.getAmount(), itemInHand.getDurability(), null);
     }
 }
