+++ bukkit/src/main/java/com/viaversion/viaversion/bukkit/platform/BukkitViaLoader.java
@@ -23,16 +23,16 @@ import com.viaversion.viaversion.api.connection.UserConnection;
 import com.viaversion.viaversion.api.minecraft.item.Item;
 import com.viaversion.viaversion.api.platform.ViaPlatformLoader;
 import com.viaversion.viaversion.api.protocol.version.ProtocolVersion;
-import com.viaversion.viaversion.bukkit.listeners.UpdateListener;
+import com.viaversion.viaversion.bukkit.listeners.base.UpdateListener;
 import com.viaversion.viaversion.bukkit.listeners.multiversion.PlayerSneakListener;
 import com.viaversion.viaversion.bukkit.listeners.v1_14_4to1_15.EntityToggleGlideListener;
 import com.viaversion.viaversion.bukkit.listeners.v1_19_3to1_19_4.ArmorToggleListener;
 import com.viaversion.viaversion.bukkit.listeners.v1_18_2to1_19.BlockBreakListener;
-import com.viaversion.viaversion.bukkit.listeners.protocol1_9to1_8.ArmorListener;
-import com.viaversion.viaversion.bukkit.listeners.protocol1_9to1_8.BlockListener;
-import com.viaversion.viaversion.bukkit.listeners.protocol1_9to1_8.DeathListener;
-import com.viaversion.viaversion.bukkit.listeners.protocol1_9to1_8.HandItemCache;
-import com.viaversion.viaversion.bukkit.listeners.protocol1_9to1_8.PaperPatch;
+import com.viaversion.viaversion.bukkit.listeners.v1_8to1_9.ArmorListener;
+import com.viaversion.viaversion.bukkit.listeners.v1_8to1_9.BlockListener;
+import com.viaversion.viaversion.bukkit.listeners.v1_8to1_9.DeathListener;
+import com.viaversion.viaversion.bukkit.listeners.v1_8to1_9.HandItemCache;
+import com.viaversion.viaversion.bukkit.listeners.v1_8to1_9.PaperPatch;
 import com.viaversion.viaversion.bukkit.listeners.v1_20_5to1_21.PaperPlayerChangeItemListener;
 import com.viaversion.viaversion.bukkit.listeners.v1_20_5to1_21.PlayerChangeItemListener;
 import com.viaversion.viaversion.bukkit.providers.BukkitAckSequenceProvider;
