+++ bukkit/src/main/java/com/viaversion/viaversion/bukkit/tasks/v1_18_2to1_19/AckSequenceTask.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.viaversion.viaversion.bukkit.tasks.protocol1_19to1_18_2;
+package com.viaversion.viaversion.bukkit.tasks.v1_18_2to1_19;
 
 import com.viaversion.viaversion.api.Via;
 import com.viaversion.viaversion.api.connection.UserConnection;
