+++ bukkit/src/main/java/com/viaversion/viaversion/bukkit/util/CollisionChecker.java
