+++ bukkit/src/main/java/com/viaversion/viaversion/bukkit/util/NMSUtil.java
